Add files via upload
intensive restructuring of code, save and load of contents is currently not possible, but the containers and the element type as 
well as the coloring will be saved and loaded.  fixed some minor issues, inserted a new function to set element drag level and
save interval of canvas surface. there is a new reset button as well. 

optimized info, license, text and notes element to activate scrollbars depending on YS UI window size. info element got some
page buttons.