Add files via upload
finished color element, it is now possible to set colors on custom buttons and rebind them at any time. 
added right click choice element, it replaces the right click program quit. right click anywhere on the
canvas(not an element on the canvas) to create the element. it offers a shutdown and a quit button.
this element will not be saved on quit. be careful, quit hasn't a dialog and just executes. shutdown
comes with a dialog.